Updated TipCalculator to use the decimal data type for numerical values.  Updated associated viewmodel and unit tests.

+++ tipclac_core/Interfaces/ITipCalculator.cs
@@ -2,10 +2,10 @@
 {
     public interface ITipCalculator
     {
-        double Total { get; set; }
-        double Tip { get; set; }
-        double TipPercent { get; set; }
-        double GrandTotal { get; }
+        decimal Total { get; set; }
+        decimal Tip { get; set; }
+        decimal TipPercent { get; set; }
+        decimal GrandTotal { get; }
 
         void CalcTip();
